'finish_project'

+++ GUI/GUI/frmKhachHang.cs
@@ -17,10 +17,10 @@ namespace QuanLyCuaHangDienThoai
         {
             InitializeComponent();
         }
-
+        List<KhachHang> khAction = new List<KhachHang>();
         BLL_KhachHang bllkh = new BLL_KhachHang();
         KhachHang kh = new KhachHang();
-        bool flagCheck; int pos = -1;
+        bool flagCheck; int pos = -1; bool undoingDelete = false; bool undoingUpdate = false;
         private void frmKhachHang_Load(object sender, EventArgs e)
         {
             LoadKhacHang();
@@ -81,11 +81,11 @@ namespace QuanLyCuaHangDienThoai
                 dgKhachHang.Columns[0].DefaultCellStyle.Alignment = DataGridViewContentAlignment.MiddleCenter;
                 dgKhachHang.Columns[0].Width = 100;
                 dgKhachHang.Columns[1].HeaderText = "Tên KH";
-                dgKhachHang.Columns[1].Width = 300;
+                dgKhachHang.Columns[1].Width = 150;
                 dgKhachHang.Columns[2].HeaderText = "Địa Chỉ";
-                dgKhachHang.Columns[2].Width = 300;
+                dgKhachHang.Columns[2].Width = 150;
                 dgKhachHang.Columns[3].HeaderText = "Điện Thoại";
-                dgKhachHang.Columns[3].Width = 300;
+                dgKhachHang.Columns[3].Width = 150;
             }
             
         }
@@ -107,6 +107,8 @@ namespace QuanLyCuaHangDienThoai
             EnbleElements();
             txtMaKH.Focus();
             flagCheck = true;
+            undoingDelete = false;
+           
         }
 
         private void btnLuu_Click(object sender, EventArgs e)
@@ -125,47 +127,83 @@ namespace QuanLyCuaHangDienThoai
             {
                 MessageBox.Show("Hãy nhập tên khách hàng", "Thông báo"); return;
             }
+            KhachHang kh = new KhachHang();
             kh.MaKH = txtMaKH.Text;
             kh.TenKH = txtTenKH.Text;
             kh.DiaChiKH = txtDiaChi.Text;
             kh.SDTKH = txtDienThoai.Text;
-            if (flagCheck)
+            if (flagCheck == true && !undoingDelete)
             {
+               
+                    if (bllkh.GetData("where makh = '" + kh.MaKH + "'") != null)
+                    {
+                        MessageBox.Show("Khách Hàng đã tồn tại", "Thông báo"); return;
+                    }
+                    else
+                    {
+                        try
+                        {
+                            kh.action = 1;
+                            khAction.Add(kh);
+                            bllkh.AddData(kh);
+                            MessageBox.Show("Thêm thành công ", "Thông báo");
+                            btnHuy.PerformClick();
+                        }
+                        catch (Exception)
+                        {
 
-                if (bllkh.GetData("where makh = '"+kh.MaKH+"'") != null)
-                {
-                    MessageBox.Show("Khách Hàng đã tồn tại", "Thông báo"); return;
-                }
-                else
-                {
-                    try
-                    {
-                        bllkh.AddData(kh);
-                        MessageBox.Show("Thêm thành công ", "Thông báo");
-                        btnHuy.PerformClick();
+                            MessageBox.Show("Lỗi! Không thêm được", "Thông báo");
+                        }
                     }
-                    catch (Exception)
-                    {
-                        
-                        MessageBox.Show("Lỗi! Không thêm được", "Thông báo");
-                    }
-                }
+               
                 
             }
-            else
+            else if (undoingDelete)
             {
                 try
                 {
-                    bllkh.EditData(kh);
-                    MessageBox.Show("Sửa thành công ", "Thông báo");
+                    kh.action = 1;
+                    khAction.Add(kh);
+                    bllkh.AddData(kh);
                     btnHuy.PerformClick();
                 }
                 catch (Exception)
                 {
 
-                    MessageBox.Show("Lỗi! Không sửa được", "Thông báo");
+                    MessageBox.Show("Undo Fail !", "Thông báo");
                 }
-                
+            }
+            else if(flagCheck == false && !undoingUpdate)
+            {    
+              
+                    try
+                    {
+                        if (pos != -1)
+                        {
+                            DataRow row = (dgKhachHang.Rows[pos].DataBoundItem as DataRowView).Row;
+                            KhachHang kh1 = new KhachHang();
+                            kh1.MaKH = row[0].ToString();
+                            kh1.TenKH = row[1].ToString();
+                            kh1.DiaChiKH = row[2].ToString();
+                            kh1.SDTKH = row[3].ToString();
+                            kh1.action = 3;
+                            khAction.Add(kh1);
+                        }   
+                        bllkh.EditData(kh);
+                        MessageBox.Show("Sửa thành công ", "Thông báo");
+                        btnHuy.PerformClick();
+                    }
+                    catch (Exception)
+                    {
+
+                        MessageBox.Show("Lỗi! Không sửa được", "Thông báo");
+                    }  
+            }
+            else if (undoingUpdate)
+            {
+                    bllkh.EditData(kh);
+                    btnHuy.PerformClick();
+               
             }
             pos = -1;
             LoadKhacHang();
@@ -177,6 +215,7 @@ namespace QuanLyCuaHangDienThoai
             btnLuu.Enabled = true;
             flagCheck = false;
             EnbleElements();
+            undoingUpdate = false;
             txtMaKH.Enabled = false;
         }
 
@@ -191,22 +230,52 @@ namespace QuanLyCuaHangDienThoai
         private void btxoa_Click(object sender, EventArgs e)
         {
             AfterClickCell();
-            DataRow row = (dgKhachHang.Rows[pos].DataBoundItem as DataRowView).Row;
-            kh.MaKH = row[0].ToString();
-            try
-            {
+            KhachHang kh = new KhachHang();
+            if(pos != -1){
+                DataRow row = (dgKhachHang.Rows[pos].DataBoundItem as DataRowView).Row;
 
-                bllkh.DeleteData(kh);
-                MessageBox.Show("Xóa Thành Công", "Thông báo");
-                LoadKhacHang();
-                pos = -1;
-                Resettext();
-                DisableElemnts();
+               
+                kh.MaKH = row[0].ToString();
+                kh.TenKH = row[1].ToString();
+                kh.DiaChiKH = row[2].ToString();
+                kh.SDTKH = row[3].ToString();
+                try
+                {
+                    kh.action = 2;   
+                    khAction.Add(kh);
+                    bllkh.DeleteData(kh);
+                    MessageBox.Show("Xóa Thành Công", "Thông báo");
+                    LoadKhacHang();
+                    pos = -1;
+                    Resettext();
+                    DisableElemnts();
+                }
+                catch (Exception)
+                {
+
+                    MessageBox.Show("Lỗi! Không thể xóa", "Thông báo");
+                }
+
             }
-            catch (Exception)
+            if (txtMaKH.Text != "")
             {
 
-                MessageBox.Show("Lỗi! Không thể xóa", "Thông báo");
+                kh.MaKH = txtMaKH.Text;
+                try
+                {
+
+                    bllkh.DeleteData(kh);
+                   
+                    LoadKhacHang();
+                    Resettext();
+                    DisableElemnts();
+                }
+                catch (Exception)
+                {
+
+                    MessageBox.Show("Undo fail!", "Thông báo");
+                }
+
             }
         }
 
@@ -214,12 +283,22 @@ namespace QuanLyCuaHangDienThoai
         {
             if (txtTimKiem.Text == "")
             {
-                MessageBox.Show("Hãy nhập tên sản phẩm cần tìm", "Thông báo"); return;
+                MessageBox.Show("Hãy nhập thông tin khách hàng cần tìm", "Thông báo"); return;
             }
+            DataTable dt = null;
             try
             {
                 kh.TenKH = txtTimKiem.Text.Trim();
-                DataTable dt = bllkh.GetData("where tenkh like N'%" + kh.TenKH + "%'");
+                if (radMaKH.Checked)
+                {
+                    dt = bllkh.GetData("where makh like N'%" + kh.TenKH + "%'");
+                }
+
+                if (radTenKH.Checked)
+                {
+                    dt = bllkh.GetData("where tenkh like N'%" + kh.TenKH + "%'");
+                }
+                
                 if (dt == null)
                 {
                     MessageBox.Show("Không tìm thấy khách hàng", "Thông báo"); return;
@@ -232,6 +311,55 @@ namespace QuanLyCuaHangDienThoai
                 throw;
             }
             btnHuy.Enabled = true;
+        }
+
+        private void btnUndo_Click(object sender, EventArgs e)
+        {
+           
+            if (khAction.Count > 0)
+            {
+                MessageBox.Show(khAction[khAction.Count - 1].action+"");
+               
+                if ( khAction[khAction.Count - 1].action == 1)
+                {
+                    txtMaKH.Text = khAction[khAction.Count - 1].MaKH;
+                    khAction.Remove(khAction[khAction.Count - 1]);
+                    btnXoa.Enabled = true;
+                    btnXoa.PerformClick();
+                }
+                else if (khAction[khAction.Count - 1].action == 2)
+                {
+                    undoingDelete = true;
+                    undoingUpdate = false;
+                    var kh = khAction[khAction.Count - 1];
+                    txtMaKH.Text = kh.MaKH;
+                    txtDiaChi.Text = kh.DiaChiKH;
+                    txtDienThoai.Text = kh.SDTKH;
+                    txtTenKH.Text = kh.TenKH;
+                    khAction.Remove(khAction[khAction.Count - 1]);
+                    btnLuu.Enabled = true;
+                    btnLuu.PerformClick();
+                }
+                else if (khAction[khAction.Count - 1].action == 3)
+                {
+                    undoingDelete = false;
+                    undoingUpdate = true;
+                    var kh = khAction[khAction.Count - 1];
+                    txtMaKH.Text = kh.MaKH;
+                    txtDiaChi.Text = kh.DiaChiKH;
+                    txtDienThoai.Text = kh.SDTKH;
+                    txtTenKH.Text = kh.TenKH;
+                    khAction.Remove(khAction[khAction.Count - 1]);
+                    btnLuu.Enabled = true;
+                    btnLuu.PerformClick();
+                }
+            }
+            
+        }
+
+        private void panel2_Paint(object sender, PaintEventArgs e)
+        {
+
         }
     }
 }
